Super-minimal skeleton that gets us feeding a file into an Engine.
I'm worried about how to structure the input subsystem to work in a browser
when we're possibly bouncing around the insides of a .tar.gz file, but I'll
deal with that later. I don't think it'll actually take much effort to
restructure that bit if I decide I need to.