Fix up a corner case for if statement parsing
Microtype does something like this:

  \ifcase 0\iftrue1\fi\else hello \fi

Where parse_int tries to expand the "\iftrue" through "\else", yielding a
complaint about expanding a bare else since there's at that point the
conditional_stack hasn't been marked. TeX has some futzing to insert an extra
\relax to stop the integer parsing. We do the same thing using a temporary
conditional_stack marker indicating that we're in the midst of parsing a
condition for an \if statement.