src/token.js: add a "pure command" token type.
This is a synthetic token that maps to a command but doesn't have a known
associated control sequence. This allows us to, e.g., safely insert a "relax"
command even if \relax has been redefined, which I believe is the correct
behavior. It also solves a proximate problem of "\the<fontcommand>" insertion
(see next commit). This feels hacky but I think it's necessary.