Fix nesting of box operations.
We would sometimes scan a box with a boxop_stack item active and sometimes
not, planning to accumulate the box by default. But of course this doesn't
work with nested box creation, because if something is being accumulated
inside a boxop, it was just peeking at the boxop_stack and finishing the boxop
prematurely. The solution, of course, is to make "accumulate" an explicit
boxop. This takes some rearrangement but is pretty straightforward.

After this fix, in the minimal LaTeX example we invoke \shipout{} with
assembled box lists that seem reasonable.