Fix Engine.ensure_horizontal()
Imagine the following situation:

  (engine is in vertical mode)
  \unhbox0

The engine reads the \unhbox and invokes it. The first thing that the command
does is ensure that the engine is in vertical mode. In this case, it starts a
new paragraph, which invokes the page builder. If \everypar is nonempty, these
tokens will get inserted at the end of the page build process. These tokens
get interposed between the \unhbox and the 0, and things break.

This explains TeX's behavior when \unhbox is encountered in vertical mode: it
does a back_input followed by new_graf. This effectively ignores the \unhbox
when it's encountered, exercises the new_graf machinery, then rereads the
\unhbox *after* the \everypar toks, with its subsequent 0 in the right place.

We modify ensure_horizontal() to make this kind of approach easily possible.
It takes the current command as an argument, so it can be requeued, and
returns true if that happens, so that the caller can know to exit early and
wait until the paragraph has been started before doing anything.