src/inputstack.js: mechanism for "cleanup" commands when an input is finished
This is to support my planned system for allowing patching of LaTeX files. I
really don't want to *actually* be patching them, so I'm working on a system
to potentially fake-\input a patch file immediately after any arbitrary LaTeX
file is loaded. I think the most robust way to do this is with hooks that
are executed by the inputstack.