src/node-io.js: another layer needed for patching
I want to be able to patch the fundamental format files latex.ltx and
plain.tex when processing them with dump-format.js. I've already set it up so
that they're processed with the \input mechanism now, so that the Engine's new
patching hooks can be invoked. But we process these files using the minimal
bundle Zip file, which doesn't contain the patch files!

I could rebuild the bundle to include the patches, but the patches are going
to be maintained in version control and it seems unwise/annoying to always
have to be rebuilding the minimal bundle. So here's a new IO stack layer that
just reads files off of the filesystem. I can have dump-format.js insert this
as a shim to get access to the patch files when processing the minimal bundle.