texpatches/tl2013/latex.ltx.post: fix up now that this is actually used
Now that the format dumper can actually apply the patch file, I realized that
it did a couple of things not-fully-correctly. I believe this version does
them right. I've tested it with a LaTeX file that imports a nonexistent
package and we now get a proper exception thrown rather than LaTeX exercising
weird edges of TeX's I/O system.