Use Latin Modern Roman fonts rather than TeX Gyre Pagella.
Pagella is much nicer, but it turns out that the only way to do math layout
with it is using the "unicode-math" package which involves a whole pile of
experimental stuff. Latin Modern Roman supports math in the traditional way.
My understand is that unicode-math is not something that we can just drop in
on an arbitrary TeX file so it's not a workable solution in the general case.