src/engine.js: another bugfix for Valref.get() calls
This fixes the case of when we tried to scan_glue() for a Dimen value that
could only be parsed by scanning an argument. I.e., \skip7=\count4. We
originally called (effectively) as_valref().get(), which would eat the '4'
token in the example above. scan_glue() would then see that the resulting
value was not a glue, and so it would delegate to scan_dimen(). This would see
the \count and repeat the same call sequence, except now as_valref().get()
would want to scan the '4' all over again, and that token has been lost. To be
able to put the token back, you'd have to have some kind of Valref.nevermind()
command that would mark that you decided you didn't want to scan that value
after all. That kind of model sounds like something that'd be a nightmare to
use consistently.