src/engine.js: err ... pay attention to results from running \output!
I'm not sure if we should really raise an exception on NeedMoreData, but I think
we might need to, and it's the safe option. Really working this out may depend
on me figuring out whether the current step() model really works or whether
we need to move to Web Workers or something. I'm deferring figuring that out for
as long as I can.