Further work attempting to safety-ize incremental parsing.
This adds a "roadblock" system to the Engine that tells it to return
NeedMoreData until something useful has come in. I added this because of a
gross problem where a font was being loaded inside an \output{} routine
and the lack of font data wasn't being caught due to the problem fixed in
the previous commit.

Now that the previous issue is fixed, I'm not sure if this infrastructure is
needed anymore. But having Fonts throw NeedMoreData in the middle of various
complex operations is definitely scary for the viability of the current
stepwise-parsing system. Again, I need to really figure out what's correct
overall, but this approach adds a bit more safety ... now that I've fixed
iterating inside the page builder.