Temporarily disable \output{}
I want to more progress with parsing a real, complex document before dealing
with all of the scary stuff that comes with invoking \output{}.

As best I can think things through, this should probably be fine. In fact, it
may be the case that the final Webtex might be able to just get away with
*never* invoking \output{} -- the use cases that I can think of generally
involve things that we don't care about in web output. However, I think it's
quite plausible that tricky LaTeX packages use temporary output routines to do
who knows what kinds of scary things, so I doubt that I'll be able to avoid it
forever. But let's see how far we can get without it.