Get the node.js variant working again.
I tried the 'webworker' and 'webworker-threads' node.js modules, and neither
worked: the former isn't compatible with Node >=0.6, apparently, and the
latter quickly ran into deadlock issues, and didn't seem to be supported.

Therefore the strategy for the node.js variant is to just do synchronous I/O
and block the main loop a lot. I'm not happy about this but it works so far.
(I'm not happy about making everything depend on synchronous I/O, either, but
it does seem to be necessary without massive interventions.)