dump-format.js: update for new Node approach
With this, I think the synchronous-I/O model is back up to compatibility with
the asynch-I/O model. However, we've lost the ability to do \shipout and
friends, so we're not at feature parity yet. That will require some design
work to figure out the best API since I think the chrome half will want to
know what's going on in both the "master" and "worker" halves of of the
browser implementation.

Before that work, however, I think I want to spend some time cleaning up code
since things are getting very ratty at this point.