Fix setting of the overall alignment width
We had several mistakes: computed the wrong glue_set value in set_glue();
inserted bare Glues into the box lists rather than BoxGlues (not noticed
because we were manually push()ing onto Box.list, so there's no way to
check types); and we missed the tabskip glues when re-creating the
contents of each row list.

After this, the \halign in the browser rendition looks a lot like the one that
pdflatex creates.