Get it so we can parse aligned math displays.
Ugh, this was a pain. There's a bunch of hacky stuff in the TeX engine here,
and I discovered a lot of bugs in my existing implementations. I also
discovered that the CanvasBox approach won't work, since amsmath pokes into
alignments in ways that utterly break if you wrap them in additional boxes.
I've changed my approach somewhat, but to get this right I think I need to add
some sort of "begin-canvas"/"end-canvas" items in the HTML renderer since we
can't guarantee that a user-facing alignment will ever get wrapped in a single
nice box that we can canvas-size.