test/tex/latex-minimal.tex: new problem: symbols
It turns out that HTML Canvas text rendering will only take Unicode code points,
whereas TeX spits out ordinals in its fonts with various funky mappings.
Obviously this is a problem that many other people have encountered but I'm not
quite sure what to do about it. pdf.js includes a complete complex font parser
and renderer so I believe I can use that to load up and draw the fonts myself
without having to worry about Unicode mapping. Alternatively, I may be able to
pre-compile mappings between ordinals and unicode points, but I'm not sure if
all of LaTeX's obscure math symbol characters are expressible in Unicode. I
wanted to do this latter with FontConfig, but it remaps the low-level glyph
indices in a way that seems to be unpredictable, which is freakin' lame.