Reduce to using an enumerated list of built-in fonts
It turns out that the compiled fonts were quite large: like 170 MB large. I'm
changing the compiler to only emit an enumerated subset of font files.

The output is still very large as these things go: 2.7 MB. So I guess I'll
have to compile fonts on the fly -- which is also true because I don't think
you can reliably count on only a small number of fonts being used in most
documents. But I'll deal with all of that later.