Get math symbol rendering kind of working!
But something weird is happening. Sizes are erratic and the last character of
each canvasbox comes out looking much nicer than the previous ones.

I guess I'll need to transmit and honor non-standard font sizes if they're
playing a role in this stuff. I suppose that may well be the case.