src/worker-api.js: add multi-pass processing here too
With this we can get bibtex references to show up decently in the HTML
incarnation. There's space getting introduced between "(" and the years in
references that probably comes from my not-too-bright algorithm noticing a
BoxGlue in the layout.

I'm hardcoding 3 iterations because I'm a bad person. This will come back to
bite me but I really don't want to spend time figuring out the logic to get
that right, at the moment.