Make the HTML render results extractable in Node.js.
There's nothing browser-specific about them because we can't make DOM objects
in Web Workers anyway. So this JSON-type output is kind of a compiled form of
the TeX document, really. Which might be a nice thing to have as an
intermediate format that we can render quickly rather than having to wait for
the LaTeX to compile every time.