src/type1-font.js: initial clone
This duplicates various bits currently living in font-compiler/, extracting
out only the bits that are necessary for parsing and rendering Type 1 fonts,
as far as I can tell. This is leading up to moving the font compiling into the
master JS file so that we don't need to enumerate and precompile the fonts
that we use.