src/type1-font.js: get it working
We now have a Type1Font class that uses Mozilla's code to parse the font data
and and general compiled JS functions for rendering on the fly. The structure
is still very awkward since the code hasn't been refactored to fit our
use-case.

The master/chrome/renderer code now needs to load up glyph-encoding.json
before it can render anything. We could also compile the data into the master
-- which increases the coupling between the master JS file and the bundle, but
perhaps I just need to come to grips with the fact that they're going to be
pretty tightly coupled.