src/math.js: fix sub_box mode, I think?
Argh, this is confusing. I ran into problems with sub_boxes when working on
accents and it looks like this is the correct solution.

Here's what I know:

- A sub-box can be either a vbox or an hbox. TTP 681.
- The new_hlist field should definitely be an hlist. I mean, it's right
  there in the name! And hpack() can get called on it and that is just
  not the right thing if it's really a vertical list.

The key is the behavior of the code in TTP 754. When a nucleus is a sub-box,
new_hlist is set directly from its "info". From the considerations above and
my best parse of the code, the change introduced in this patch is what it's
doing, rather than what I had before.

Unfortunately this changeseems to introduce more levels of unnecessary hbox
wrapping that I believe need to go away at some point. A \showlists after
$\hat \alpha$ yields

    \mathon
    \vbox(6.94444+0.0)x6.43404
    .\hbox(6.94444+0.0)x0.0, shifted 0.9948
    ..\OT1/cmr/m/n/10 ^
    .\kern-4.30554
    .\hbox(4.30554+0.0)x6.43404
    ..\OML/cmm/m/it/10

    \mathoff

while I'm currently returning output of that, wrapped in a 1-item hbox.