Fix up implementation of \eTeXversion
\eTeXversion is a read-only integer. Because it can't be set, there's no point
in hooking it up to the eqtb nesting machinery and all that, so it seems
better to implement it as a command with a constantvalref. I looked into
making it possible to define read-only named parameters, but there are very
few of them and, again, all of the nesting machinery is useless.

Anyway, there was a hangover of it being defined as a named parameter rather
than a command, so the right implementation wasn't being used.