Add a system for suppressing automatic <p> tags
It turns out that begin_graf() gets called in a lot of places where we don't
want to insert a <p> tag; i.e, we get <p>/</p> nested inside <h1> tags. That's
not so nice. This new system adds a stack that allows code to turn automatic
<p> insertion on or off. This is the first approach that I came up with; it
seems to work, but it might need tweaking.