Fix subtleties in box voiding
This fixes the mysterious issue with the ever-expanding list of <a> tags in
the References section! Yay!

The basic issue is that the interaction between the nesting of box registers
and the way that TeX voids them is not what you might think. For instance:

  \setbox0=\hbox{yo}
  { \box0 }
  \showbox0

Now, the bare \box0 command will void out box#0, but you might think that the
outermost \box0 will be preserved because the \box0 command ocurred inside a
group. However, as described in the code comments in the commit, that's not
how things are done -- which does make sense in retrospect, because otherwise
TeX would have to copy box contents for operations like \raise and the whole
point is that it doesn't want to.