Add "delimiter group" HTML tags and use them for reference text.
The idea here is to make it possible to properly tag constructions like
\begin{enumerate} \item ... \item ... \end{enumerate}. Now, the text syntax
HTML5 doesn't require end-tags for <li>s in these situations, which would make
things pretty easy. But our approach to tagging requires explicit end tags,
which makes things annoying because (1) each "\item" equivalent needs to know
if there's a prior <li> equivalent to close, and (2) each "\end{enumerate}"
equivalent needs to know the same thing.

Our approach is to add two variants to the "start-tag" and "end-tag" specials
called "delim-tag" and "end-delim-tag". The "delim-tag" special always inserts
a start tag, but also closes the preceding tag if appropriate. The
"end-delim-tag" special just closes the preceding tag if appropriate. Tags
delimited in this way are assigned "delimiter group" identifiers for some
sanity-checking in case we might end up nesting instances of this construct.
We don't just have a single "maybe end previous tag" construct for the same
reason; without a "start delimited tag" construct, there'd be no way to be
sure that we were ending a tag that was indeed supposed to be delimited in
this way.

We use this construct to delimit \bibitems in special <div> tags. The goal
here is to eventually add an effect where the reference info pops up when a
reference link is moused over.